+++ tapo-py/examples/tapo_p110.py
@@ -28,11 +28,11 @@ async def main():
     device_info = await device.get_device_info()
     print(f"Device info: {device_info.to_dict()}")
 
-    device_usage = await device.get_device_usage()
-    print(f"Device usage: {device_usage.to_dict()}")
-
     current_power = await device.get_current_power()
     print(f"Current power: {current_power.to_dict()}")
+
+    device_usage = await device.get_device_usage()
+    print(f"Device usage: {device_usage.to_dict()}")
 
     energy_usage = await device.get_energy_usage()
     print(f"Energy usage: {energy_usage.to_dict()}")